+++ components/App.jsx
@@ -1,42 +1,34 @@
 var React = require('react');
 var Router = require('react-router');
 var Reflux = require('reflux');
-var AppLeftNav = require('./common/AppLeftNav');
-var { Route, DefaultRoute, RouteHandler, Link } = Router;
+var Sidebar = require('./layout/Sidebar2');
+var Header = require('./layout/Header');
+var { RouteHandler } = Router;
 
-var mui = require('material-ui');
-var { AppBar, AppCanvas, Menu, IconButton } = mui;
-var ThemeManager = new mui.Styles.ThemeManager();
-
-console.log('palette', ThemeManager.palette);
-//console.log('AppBar', AppBar.getStyle());
-var customTheme = require('../stylesheets/theme');
-
-var flux = require('../scripts/app');
-
-ThemeManager.setTheme(customTheme);
+var store = require('../scripts/store');
+var actions = require('../scripts/actions');
 
 var App = React.createClass({
 
   mixins: [Reflux.ListenerMixin],
 
-  // material-ui
-  childContextTypes: {
-    muiTheme: React.PropTypes.object
-  },
-  getChildContext: function() {
-    return {
-      muiTheme: ThemeManager.getCurrentTheme()
-    };
-  },
-
   componentDidMount: function() {
     //Listen to any change from the store (@trigger() in the store)
-    this.listenTo(flux.store, this.onChangeStore);
+    this.listenTo(store, this.onChangeStore);
+    var data = this.props.data;
+    actions.getProjects.completed(data);
+
+    //Remove the splash screen by removing the .nojs class
+    var elements = document.querySelectorAll('.nojs');
+    Array.prototype.forEach.call( elements, (el) => el.classList.remove('nojs'));
+
+    //Add the stylesheets to overwrite inline styles defined in index.html
+    require('./layout/layout.styl');
+    require('../stylesheets/base.styl');
+    require('../stylesheets/table.styl');
   },
   getInitialState: function() {
-    console.log(flux.store.getInitialState());
-    return flux.store.getInitialState();
+    return store.getInitialState();
   },
   onChangeStore: function(storeData) {
     //Store has changed => update the view.
@@ -46,22 +38,19 @@ var App = React.createClass({
   render: function() {
     console.log('Render the top level component.', this.state);
     return (
-      <AppCanvas predefinedLayout={1}>
-        <AppBar
-           title={ 'bestof.js.org' }
-           className="mui-dark-theme"
-           onLeftIconButtonTouchTap={this._onLeftIconButtonTouchTap}
-           onMenuIconButtonTouchTap={this._onLeftIconButtonTouchTap}
-           zDepth={0}
-        >
-        </AppBar>
+      <div id="layout">
 
-        <AppLeftNav
-          ref="leftNav"
-          tags={ this.state.tags }
+        <Sidebar
+          tags={ this.state.tags}
+          selectedTag={ this.state.selectedTag }
         />
 
-        <div className="container">
+        <div id="main">
+
+          <Header
+            searchText={this.state.searchText}
+          />
+
           <RouteHandler
             allProjects={ this.state.allProjects }
             searchText={this.state.searchText}
@@ -74,27 +63,15 @@ var App = React.createClass({
             selectedSort={ this.state.selectedSort }
             project={ this.state.project }
             tag={ this.state.tag }
+            errorMessage={ this.state.errorMessage }
           />
+
         </div>
 
-      </AppCanvas>
+      </div>
     );
   },
 
-  _onLeftIconButtonTouchTap: function() {
-    console.log('tap!');
-    this.refs.leftNav.toggle();
-  },
-
-
 });
-
-App.contextTypes = {
-  router: React.PropTypes.func
-};
-
-App.childContextTypes = {
-  muiTheme: React.PropTypes.object
-};
 
 module.exports = App;