+++ components/home/SearchResultList.jsx
@@ -1,34 +1,56 @@
 var React = require('react');
+var Router = require('react-router');
+var {Link} = Router;
 
 var ProjectList = require('../projects/ProjectList');
 var SearchText = require('../common/utils/SearchText');
+var TagLabel = require('../tags/TagTitle');
 
 var SearchResultList = React.createClass({
 
   render: function() {
     var projects = this.props.projects;
+    var tag = this.props.tag;
     var style={
       container: {
-        margin: '20px 0'
+        margin: '0'
       }
     };
     return (
       <div style={ style.container }>
-        { projects.length > 0 ? (
-          <div>
-            <h3>
-              Results for <SearchText>{ this.props.searchText }</SearchText>:{' '}
-              {projects.length } projects found.
-              </h3>
-            <ProjectList
-              projects = { projects }
-              maxStars = { projects[0].stars}
-              showDescription = { true }
-            />
+
+        { tag.name && (
+          <div style={{ fontSize: 18, marginBottom: 20 }}>
+            <Link to="home">
+              <TagLabel tag={ tag } onClickXXX={ this.props.onRemoveTag }/>
+            </Link>
+            <span style={{  marginLeft: 10 }}>{projects.length } projects</span>
           </div>
-        ) : (
+        ) }
+
+        { (this.props.searchText && projects.length > 0) && (
+          <h3>
+            Results for <SearchText>{ this.props.searchText }</SearchText>:{' '}
+            {projects.length } projects found.
+           </h3>
+        )}
+
+        { (this.props.searchText && projects.length === 0) && (
           <div>No project found for <SearchText>{ this.props.searchText }</SearchText></div>
         )}
+
+
+        { projects.length > 0 && (
+         <div>
+           <ProjectList
+             projects = { projects }
+             maxStars = { projects[0].stars}
+             showDescription = { true }
+             showURL = { true }
+           />
+         </div>
+       ) }
+
     </div>
     );
   }