+++ webpack.local.config.js
@@ -20,7 +20,7 @@ module.exports = {
     app: [
       "webpack-dev-server/client?http://localhost:8080",
       "webpack/hot/only-dev-server",
-      "./entry.jsx"
+      "./components/entry.jsx"
     ],
     vendor: [
       "react",